+++ java/arcs/core/entity/EntityBase.kt
@@ -26,8 +26,8 @@ import arcs.core.data.util.ReferencableList
 import arcs.core.data.util.ReferencablePrimitive
 import arcs.core.data.util.toReferencable
 import arcs.core.storage.Reference as StorageReference
+import arcs.core.util.BigInt
 import arcs.core.util.Time
-import java.math.BigInteger
 import kotlin.reflect.KProperty
 
 open class EntityBase(
@@ -196,7 +196,7 @@ open class EntityBase(
                 PrimitiveType.Double -> require(value is Double) {
                     "Expected Double for $context$entityClassName.$field, but received $value."
                 }
-                PrimitiveType.BigInt -> require(value is BigInteger) {
+                PrimitiveType.BigInt -> require(value is BigInt) {
                     "Expected BigInt for $context$entityClassName.$field, but received $value."
                 }
             }
@@ -392,7 +392,7 @@ private fun toReferencable(value: Any, type: FieldType): Referencable = when (ty
         PrimitiveType.Char -> (value as Char).toReferencable()
         PrimitiveType.Float -> (value as Float).toReferencable()
         PrimitiveType.Double -> (value as Double).toReferencable()
-        PrimitiveType.BigInt -> (value as BigInteger).toReferencable()
+        PrimitiveType.BigInt -> (value as BigInt).toReferencable()
     }
     is FieldType.EntityRef -> (value as Reference<*>).toReferencable()
     // TODO(b/155025255)
